+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * fix: Fix the bug that Datex.formatTimeLength(1590L, "%ms?ms") should return '1590ms' instead of '1ms'
 * fix: Fix the bug that Datex.formatTimeLength(90L, "%MS?ms") should return '090ms' instead of '90ms'
 * fix: Fix the bug that Datex.formatTimeLength(0L, "%d?d %h?h %m?m %s?s %ms?ms") should return '0ms' instead of '0s'
+* add: Filex adds testFilterFileNameIllegalCharacters() and testFilterFileNameIllegalCharactersOrNull() method
 
 
 ## v1.0.0-rc01